+++ GlossExample/Pods/Gloss/Pod/Classes/Decoder.swift
@@ -33,9 +33,9 @@ public struct Decoder {
     /**
     Returns function to decode JSON to value type
     
-    - parameter key: JSON key used to set value
+    :parameter: key JSON key used to set value
     
-    - returns: Function decoding JSON to an optional value type
+    :returns: Function decoding JSON to an optional value type
     */
     public static func decode<T>(key: String) -> JSON -> T? {
         return {
@@ -53,9 +53,9 @@ public struct Decoder {
     Returns function to decode JSON to value type
     for objects that conform to the Glossy protocol
     
-    - parameter key: JSON key used to set value
+    :parameter: key JSON key used to set value
     
-    - returns: Function decoding JSON to an optional value type
+    :returns: Function decoding JSON to an optional value type
     */
     public static func decode<T: Decodable>(key: String) -> JSON -> T? {
         return {
@@ -74,9 +74,9 @@ public struct Decoder {
     Returns function to decode JSON to array
     of enum values
     
-    - parameter key: JSON key used to set value
+    :parameter: key JSON key used to set value
     
-    - returns: Function decoding JSON to an optional array
+    :returns: Function decoding JSON to an optional array
     */
     public static func decodeArray<T: RawRepresentable>(key: String) -> JSON -> [T]? {
         return {
@@ -102,9 +102,9 @@ public struct Decoder {
     Returns function to decode JSON to array
     for objects that conform to the Glossy protocol
     
-    - parameter key: JSON key used to set value
+    :parameter: key JSON key used to set value
     
-    - returns: Function decoding JSON to an optinal array
+    :returns: Function decoding JSON to an optinal array
     */
     public static func decodeArray<T: Decodable>(key: String) -> JSON -> [T]? {
         return {
@@ -114,7 +114,9 @@ public struct Decoder {
                 var models: [T] = []
                 
                 for subJSON in jsonArray {
-                    models.append(T(json: subJSON))
+                    if let model = T(json: subJSON) {
+                        models.append(model)
+                    }
                 }
                 
                 return models
@@ -127,10 +129,10 @@ public struct Decoder {
     /**
     Returns function to decode JSON to date
     
-    - parameter key:           JSON key used to set value
-    - parameter dateFormatter: Formatter used to format date
+    :parameter: key           JSON key used to set value
+    :parameter: dateFormatter Formatter used to format date
     
-    - returns: Function decoding JSON to an optional date
+    :returns: Function decoding JSON to an optional date
     */
     public static func decodeDate(key: String, dateFormatter: NSDateFormatter) -> JSON -> NSDate? {
         return {
@@ -147,8 +149,8 @@ public struct Decoder {
     /**
     Returns function to decode JSON to ISO8601 date
     
-    - parameter key:           JSON key used to set value
-    - parameter dateFormatter: Formatter with ISO8601 format
+    :parameter: key           JSON key used to set value
+    :parameter: dateFormatter Formatter with ISO8601 format
     
     - returns: Function decoding JSON to an optional ISO8601 date
     */
@@ -163,9 +165,9 @@ public struct Decoder {
     /**
     Returns function to decode JSON to enum value
     
-    - parameter key: JSON key used to set value
+    :parameter: key JSON key used to set value
     
-    - returns: Function decoding JSON to an optional enum value
+    :returns: Function decoding JSON to an optional enum value
     */
     public static func decodeEnum<T: RawRepresentable>(key: String) -> JSON -> T? {
         return {
@@ -182,9 +184,9 @@ public struct Decoder {
     /**
     Returns function to decode JSON to URL
     
-    - parameter key: JSON key used to set value
+    :parameter: key JSON key used to set value
     
-    - returns: Function decoding JSON to an optional URL
+    :returns: Function decoding JSON to an optional URL
     */
     public static func decodeURL(key: String) -> JSON -> NSURL? {
         return {
@@ -196,105 +198,6 @@ public struct Decoder {
             
             return nil
         }
-    }
-    
-    // MARK: - Force Decoders
-    
-    /**
-    Returns function to decode JSON to value type
-    
-    - parameter key: JSON key used to set value
-    
-    - returns: Function decoding JSON to a value type
-    */
-    public static func forceDecode<T>(key: String) -> JSON -> T {
-        return { return decode(key)($0)! }
-    }
-    
-    /**
-    Returns function to decode JSON to value type
-    for objects that conform to the Glossy protocol
-    
-    - parameter key: JSON key used to set value
-    
-    - returns: Function decoding JSON to a value type
-    */
-    public static func forceDecode<T: Decodable>(key: String) -> JSON -> T {
-        return { return decode(key)($0)! }
-    }
-    
-    /**
-    Returns function to decode JSON to array
-    of enum values
-    
-    - parameter key: JSON key used to set value
-    
-    - returns: Function decoding JSON to a array
-    */
-    public static func forceDecodeArray<T: RawRepresentable>(key: String) -> JSON -> [T] {
-        return { return decodeArray(key)($0)! }
-    }
-    
-    /**
-    Returns function to decode JSON to array
-    for objects that conform to the Glossy protocol
-    
-    - parameter key: JSON key used to set value
-    
-    - returns: Function decoding JSON to a array
-    */
-    public static func forceDecodeArray<T: Decodable>(key: String) -> JSON -> [T] {
-        return { return decodeArray(key)($0)! }
-    }
-    
-    /**
-    Returns function to decode JSON to date
-    
-    - parameter key:           JSON key used to set value
-    - parameter dateFormatter: Formatter used to format date
-    
-    - returns: Function decoding JSON to a date
-    */
-    public static func forceDecodeDate(key: String, dateFormatter: NSDateFormatter) -> JSON -> NSDate {
-        return { return decodeDate(key, dateFormatter: dateFormatter)($0)! }
-    }
-    
-    /**
-    Returns function to decode JSON to ISO8601 date
-    
-    - parameter key:           JSON key used to set value
-    - parameter dateFormatter: Formatter with ISO8601 format
-    
-    - returns: Function decoding JSON to a ISO8601 date
-    */
-    public static func forceDecodeDateISO8601(key: String) -> JSON -> NSDate {
-        let dateFormatter = NSDateFormatter()
-        dateFormatter.locale = NSLocale(localeIdentifier: "en_US_POSIX")
-        dateFormatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ssZZZZZ"
-        
-        return Decoder.forceDecodeDate(key, dateFormatter: dateFormatter)
-    }
-    
-    /**
-    Returns function to decode JSON to enum value
-    
-    - parameter key: JSON key used to set value
-    
-    - returns: Function decoding JSON to an enum value
-    */
-    public static func forceDecodeEnum<T: RawRepresentable>(key: String) -> JSON -> T {
-        return { return decodeEnum(key)($0)! }
-    }
-    
-    /**
-    Returns function to decode JSON to URL
-    
-    - parameter key: JSON key used to set value
-    
-    - returns: Function decoding JSON to a URL
-    */
-    public static func forceDecodeURL(key: String) -> JSON -> NSURL {
-        return { return decodeURL(key)($0)! }
     }
     
 }