+++ GlossExample/Pods/Target Support Files/Pods-GlossExample/Pods-GlossExample-environment.h
@@ -9,12 +9,12 @@
 // Gloss
 #define COCOAPODS_POD_AVAILABLE_Gloss
 #define COCOAPODS_VERSION_MAJOR_Gloss 0
-#define COCOAPODS_VERSION_MINOR_Gloss 4
+#define COCOAPODS_VERSION_MINOR_Gloss 5
 #define COCOAPODS_VERSION_PATCH_Gloss 0
 
 // Gloss/Extensions
 #define COCOAPODS_POD_AVAILABLE_Gloss_Extensions
 #define COCOAPODS_VERSION_MAJOR_Gloss_Extensions 0
-#define COCOAPODS_VERSION_MINOR_Gloss_Extensions 4
+#define COCOAPODS_VERSION_MINOR_Gloss_Extensions 5
 #define COCOAPODS_VERSION_PATCH_Gloss_Extensions 0